requirements.txt
RPi.GPIO
telnetlib3
flask
usb
mcculw
flask-socketio 
eventlet
pyusb
pexpect